# backend/swagger-ui/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Baustellenverwaltung",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://localhost:8080"
    }
  ],
...
  "paths": {
    "/account/login": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Request login",
        "operationId": "login",
        "requestBody": {
          "description": "Email and Password of registered Account that requests to login",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {}
          },
          "400": {
            "description": "Invalid email provided",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Account"
      }
    },
    "/account/reset": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Request account reset",
        "operationId": "resetUser",
        "requestBody": {
          "description": "Email of registered Account that requests to reset his Login information",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountReset"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {}
          },
          "400": {
            "description": "Invalid email provided",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Account"
      }
    },
    "/account/password": {
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Update user password",
        "operationId": "updatePassword",
        "requestBody": {
          "description": "Object needs to contain new Password and resetId",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountUpdatedPassword"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation"
          },
          "400": {
            "description": "Couldn't update Password",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Account"
      }
    },
    "/account/invite": {
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Invite a User",
        "operationId": "inviteUser",
        "requestBody": {
          "description": "Email of Person that you want to invite",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountInvite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "successful operation, returns invited User-Object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email provided or user already invited",
            "content": {}
          },
          "401": {
            "description": "Unauthorized",
            "content": {}
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "x-codegen-request-body-name": "body",
        "x-swagger-router-controller": "Account"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "phone": {
            "type": "number"
          },
          "address": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "invited"
            ]
          },
          "role": {
            "type": "string"
          }
        },
        "example": {
          "zip": "zip",
          "address": "address",
          "role": "role",
          "phone": 12345667,
          "city": "city",
          "name": "name",
          "language": "language",
          "id": 0,
          "email": "email",
          "status": "active"
        },
        "xml": {
          "name": "User"
        }
      },
...
      "AccountUpdatedPassword": {
        "required": [
          "password",
          "resetId"
        ],
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "resetId": {
            "type": "string"
          }
        },
        "xml": {
          "name": "AccountUpdatedPassword"
        }
      },
      "AccountInvite": {
        "required": [
          "email",
          "role",
          "name"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        },
        "example": {
          "name": "Kevin ...",
          "email": "kevin.test@....",
          "role": "admin"
        }
      },
      "AccountReset": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "xml": {
          "name": "AccountReset"
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}